update: binary-clf finished

+++ src/train.py
@@ -2,13 +2,14 @@ import yaml
 import pytorch_lightning as pl
 import pandas as pd
 
-from models import BaseModel
+from models import BaseModel, BinaryClassifier
 from loader import KLUEDataLoader
 from transformers import AutoTokenizer
 from pytorch_lightning.loggers import WandbLogger
 from pytorch_lightning.callbacks import ModelCheckpoint, EarlyStopping
 from utils import get_result_name, num_to_label, remove_pad_tokens
 from typing import Optional
+import os
 import wandb
 import shutil
 
@@ -27,13 +28,13 @@ def train(cfg, result_name :Optional[str] = None):
     tokenizer = AutoTokenizer.from_pretrained(
         cfg["model_name"], model_max_length=cfg["max_len"]
     )
-    model = BaseModel(tokenizer, cfg)
+    model = BinaryClassifier(tokenizer, cfg)
     dataloader = KLUEDataLoader(tokenizer, cfg)
 
     # wandb logger, ggul_tiger 팀으로 run이 기록됩니다.
     logger = WandbLogger(
         name=result_name, 
-        project="KLUE", 
+        project="KLUE2", 
         entity="ggul_tiger"
     )
     logger.experiment.config.update(cfg)
@@ -41,7 +42,7 @@ def train(cfg, result_name :Optional[str] = None):
 
     trainer = pl.Trainer(
         accelerator="gpu",
-        strategy="ddp",
+        # strategy="ddp",
         max_epochs=cfg["epoch"],
         logger=logger,
         default_root_dir=cfg["result_dir"] + result_name,
@@ -70,15 +71,15 @@ def train(cfg, result_name :Optional[str] = None):
     except KeyboardInterrupt:
         pass
     finally:
-        if trainer.current_epoch>=2:
+        if trainer.current_epoch>=1:
             trainer.test(model=model, datamodule=dataloader, ckpt_path="best")
             # validation data로 모델의 prediction 결과를 result 폴더에 csv파일로 저장합니다.
             val_result = model.val_result
             val_result["tokenized"] = remove_pad_tokens(
                 val_result["tokenized"], tokenizer.pad_token
             )
-            val_result["target"] = num_to_label(val_result["target"])
-            val_result["predict"] = num_to_label(val_result["predict"])
+            # val_result["target"] = num_to_label(val_result["target"])
+            # val_result["predict"] = num_to_label(val_result["predict"])
             val_result_df = pd.DataFrame(val_result)
             val_result_df.to_csv(
                 cfg["result_dir"] + result_name + "/val_result.csv", index=False
@@ -89,9 +90,9 @@ def train(cfg, result_name :Optional[str] = None):
             api = wandb.Api()
             run = api.run("ggul_tiger/KLUE/{}".format(logger.version))
             run.delete(delete_artifacts=True)
-
-            print('deleteing local folder : {}'.format(result_name))
-            shutil.rmtree('results/{}'.format(result_name))
+            if os.path.exists(result_name):
+                print('deleteing local folder : {}'.format(result_name))
+                shutil.rmtree('results/{}'.format(result_name))
 
 if __name__=="__main__":
     # load config
